Customer Adds product with PLU Code
Added methods to customer class to allow for PLU coded items to be added to cart.
Added to reciept class to allow for PLU coded items to be added to the subtotal
Process Item handler class now uses more methods in customer class to add PLU code to cart, and then notifies place in bagging area observer.